test(chat-access): drive Task 7 /send + ensureSession gates through the real access service
Remove vi.mock of chat-integration-access-service from both Task 7 test
files. Wire a real in-memory better-sqlite3 behind @shared/lib/db using
the getter pattern (same as chat-integration-access-gate.test.ts), run
migrations, and seed rows so isChatAllowed exercises genuine SQL.

x-agent-chat.test.ts: seed integration-1 as slack/no-approval in
beforeEach (all default-path tests pass); the 403 case updates the row
to telegram+require_approval=1 before the request so isChatAllowed
returns false via real DB lookup.

Rename chat-integration-manager.task7.test.ts to
chat-integration-manager-ensure-session-access.test.ts; seed the
telegram integration with require_approval=1; throw test has no access
row (denied by SQL), success test seeds status='allowed' row.
